feat(render) WIP: textures, using a fixed size texture array
currently using an 8 long texture array with seperate samplers. 

also move to a trait for resolving textures from the bsp file rather than adding them manually. currently stuck with one static implementation though. this means all textures in the file are loaded when the map is.

not currently drawing with anything other than the first texture array, so files with >8 textures will fail.